# api/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Blog API
  version: 1.0.0
  description: API documentation for the Blog application.
paths:
...
  /api/v1/comments:
    post:
      summary: Create a new comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - blog_post_id
                - user_id
                - comment
              properties:
                blog_post_id:
                  type: string
                  format: uuid
                  description: The ID of the blog post to which the comment is related
                  example: "123e4567-e89b-12d3-a456-426614174000"
                user_id:
                  type: string
                  format: uuid
                  description: The ID of the user making the comment
                  example: "123e4567-e89b-12d3-a456-426614174001"
                comment:
                  type: string
                  description: The content of the comment
                  example: "This is a comment."
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment added successfully"
                  comment:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174002"
                      blog_post_id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174000"
                      user_id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174001"
                      comment:
                        type: string
                        example: "This is a comment."
                      created_at:
                        type: string
                        format: date-time
                        example: "2024-08-25T14:15:22Z"
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required fields"
        '404':
          description: Blog post or user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Blog post not found"
        '500':
          description: Database or internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Database error occurred"
                  details:
                    type: string
                    example: "Detailed error message"
  /api/v1/blog_posts/{post_id}/comments:
    get:
      summary: Get comments for a blog post
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: post_id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the blog post
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: The page number to retrieve
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
          description: The number of comments per page
      responses:
        '200':
          description: A list of comments for the blog post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  comments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          example: "123e4567-e89b-12d3-a456-426614174002"
                        user_id:
                          type: string
                          format: uuid
                          example: "123e4567-e89b-12d3-a456-426614174001"
                        comment:
                          type: string
                          example: "This is a comment."
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-08-25T14:15:22Z"
                  page:
                    type: integer
                    example: 1
                  per_page:
                    type: integer
                    example: 10
                  total_comments:
                    type: integer
                    example: 100
                  total_pages:
                    type: integer
                    example: 10
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Page number and per_page must be positive integers"
        '404':
          description: Blog post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Blog post not found"
        '500':
          description: Database or internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Database error occurred"
                  details:
                    type: string
                    example: "Detailed error message"
  /api/v1/comments/{comment_id}:
    get:
      summary: Get a specific comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the comment to retrieve
      responses:
        '200':
          description: The requested comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174003"
                  blog_post_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  user_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174001"
                  comment:
                    type: string
                    example: "This is a comment."
                  created_at:
                    type: string
                    format: date-time
                    example: "2024-08-25T14:15:22Z"
        '400':
          description: Invalid comment ID format
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid comment ID format"
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Comment not found"
        '500':
          description: Database or internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Database error occurred"
                  details:
                    type: string
                    example: "Detailed error message"
    put:
      summary: Update a specific comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the comment to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  example: "Updated comment text"
      responses:
        '200':
          description: The comment was updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment updated successfully"
                  comment:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174003"
                      blog_post_id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174000"
                      user_id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174001"
                      comment:
                        type: string
                        example: "Updated comment text"
                      created_at:
                        type: string
                        format: date-time
                        example: "2024-08-25T14:15:22Z"
        '400':
          description: Missing comment text in the request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing comment text"
        '403':
          description: Unauthorized access to update the comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized access"
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Comment not found"
        '500':
          description: Database or internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Database error occurred"
                  details:
                    type: string
                    example: "Detailed error message"
    delete:
      summary: Delete a specific comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: comment_id
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the comment to delete
      responses:
        '200':
          description: The comment was deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment deleted successfully"
        '403':
          description: Unauthorized access to delete the comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "You are not authorized to delete this comment"
        '404':
          description: Comment or associated blog post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Comment not found"
        '500':
          description: Database or internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Database error occurred"
                  details:
                    type: string
                    example: "Detailed error message"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT